[hlsl] Remove i32 offsets for atomic builtins
HLSL does not require that the type of the offset matches the type of
the atomic. We always generate u32 offsets and have some transforms
that are assuming u32 offsets.

Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/300335
Commit-Queue: James Price <jrprice@google.com>

+++ test/tint/bug/chromium/1273230.wgsl.expected.dxc.hlsl
@@ -64,7 +64,7 @@ void doIgnore() {
   LUT.GetDimensions(v_10);
   uint v_11 = ((v_10 / 4u) - 1u);
   int v_12 = int(0);
-  LUT.InterlockedOr(int((0u + (min(uint(int(0)), v_11) * 4u))), int(0), v_12);
+  LUT.InterlockedOr((0u + (min(uint(int(0)), v_11) * 4u)), int(0), v_12);
   int g55 = v_12;
 }
 
@@ -92,7 +92,7 @@ void main_count_inner(uint3 GlobalInvocationID) {
   uint v_16 = 0u;
   LUT.GetDimensions(v_16);
   int v_17 = int(0);
-  LUT.InterlockedAdd(int((0u + (min(i1, ((v_16 / 4u) - 1u)) * 4u))), int(1), v_17);
+  LUT.InterlockedAdd((0u + (min(i1, ((v_16 / 4u) - 1u)) * 4u)), int(1), v_17);
   int triangleOffset = v_17;
 }
